+++ R/rdb_dimensions.R
@@ -38,9 +38,9 @@
 #' \code{data.table}s.
 #' @examples
 #' \dontrun{
-#' rdb_dimensions(provider_code = "IMF", dataset_code = "WEO")
+#' rdb_dimensions(provider_code = "IMF", dataset_code = "WEO:2019-10")
 #' 
-#' rdb_dimensions(provider_code = "IMF", dataset_code = "WEO", simplify = TRUE)
+#' rdb_dimensions(provider_code = "IMF", dataset_code = "WEO:2019-10", simplify = TRUE)
 #' 
 #' rdb_dimensions(provider_code = "IMF")
 #' 
@@ -50,12 +50,12 @@
 #' options(rdbnomics.progress_bar_dimensions = FALSE)
 #' 
 #' rdb_dimensions(
-#'   provider_code = "IMF", dataset_code = "WEO",
+#'   provider_code = "IMF", dataset_code = "WEO:2019-10",
 #'   use_readLines = TRUE
 #' )
 #' 
 #' rdb_dimensions(
-#'   provider_code = "IMF", dataset_code = "WEO",
+#'   provider_code = "IMF", dataset_code = "WEO:2019-10",
 #'   curl_config = list(proxy = "<proxy>", proxyport = <port>)
 #' )
 #' }